requirements.txt
Django==4.0.4
django-url-filter==0.3.15
djangorestframework==3.13.1
gunicorn==20.1.0
jupyterlab==3.3.4
psycopg2-binary==2.9.3
requests==2.27.1
